Refactoring vote screen so we can use the components in view/Results - vier/Result is not consistent w/ vote screen so I refactored the components from the vote screen that will make up view/Results     - refactored title into seperate component     - refactored the charts into seperate components     - refactored the room images into seperate components
