+++ src/components/RoomImages/index.js
@@ -1,0 +1,1 @@
+export { default as RoomImages } from "./RoomImages"
