+++ src/components/RoomTitle/index.js
@@ -1,0 +1,1 @@
+export { default as RoomTitle } from "./RoomTitle"
