+++ src/components/VotingChart/index.js
@@ -1,0 +1,1 @@
+export { default as VotingChart } from "./VotingChart"
